add unit test for db2
