+++ src/main/java/org/tron/core/db/TronStoreWithRevoking.java
@@ -25,7 +25,7 @@ import org.tron.core.exception.ItemNotFoundException;
 
 @Slf4j
 public abstract class TronStoreWithRevoking<T extends ProtoCapsule> implements ITronChainBase<T> {
-
+  @Getter // only for unit test
   protected IRevokingDB revokingDB;
   private TypeToken<T> token = new TypeToken<T>(getClass()) {};
   @Autowired
